test: standardize Rust test locations per
