+++ crates/adoc-cli/tests/support/mod.rs
@@ -1,0 +1,41 @@
+use std::fs;
+use std::path::PathBuf;
+use std::time::{SystemTime, UNIX_EPOCH};
+
+pub(crate) fn fixture_path(relative: &str) -> PathBuf {
+    PathBuf::from(env!("CARGO_MANIFEST_DIR"))
+        .join("tests")
+        .join("fixtures")
+        .join(relative)
+}
+
+pub(crate) struct TestWorkspace {
+    pub(crate) root: PathBuf,
+}
+
+impl TestWorkspace {
+    pub(crate) fn new(name: &str) -> Self {
+        let nonce = SystemTime::now()
+            .duration_since(UNIX_EPOCH)
+            .expect("clock is after epoch")
+            .as_nanos();
+        let root = std::env::temp_dir().join(format!("adoc-{name}-{nonce}"));
+        fs::create_dir_all(&root).expect("test workspace can be created");
+        Self { root }
+    }
+
+    pub(crate) fn write(&self, relative_path: &str, contents: &str) -> PathBuf {
+        let path = self.root.join(relative_path);
+        if let Some(parent) = path.parent() {
+            fs::create_dir_all(parent).expect("parent directory can be created");
+        }
+        fs::write(&path, contents).expect("test source can be written");
+        path
+    }
+}
+
+impl Drop for TestWorkspace {
+    fn drop(&mut self) {
+        let _ = fs::remove_dir_all(&self.root);
+    }
+}
